third_party/WebKit/LayoutTests/platform/mac/virtual/outofblink-cors-http/http/tests/htmlimports/cross-origin-expected.txt

PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
